+++ docs/CHANGELOG.md
@@ -6,8 +6,15 @@ All notable changes to this project will be documented in this file.
 
 
 ## [0.3.0] - 2022-12-DD
+
+### Added
 - Support for Python3.9 [#6](https://github.com/hakancelikdev/pydbm/issues/6)
+
+### Changed
 - Documentations improve, add version support [#11](https://github.com/hakancelikdev/pydbm/issues/11)
+
+### Fixed
+- Fix exception name -> BaseException [#12](https://github.com/hakancelikdev/pydbm/issues/12)
 
 ## [0.2.0] - 2022-11-29